requirements.txt
openai
langchain
chromadb
faiss-cpu
sentence-transformers
opencv-python
pytesseract
streamlit
pandas
scikit-learn
matplotlib
mlflow
python-dotenv
PyMuPDF
faker
